pyproject.toml
[project]
name = "ultra-doc-intelligence"
version = "0.1.0"
description = "Submission-focused logistics document intelligence POC"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "docling>=2.40.0",
  "fastapi>=0.115.0",
  "httpx>=0.27.0",
  "openai>=1.40.0",
  "pydantic>=2.8.0",
  "pypdf>=5.0.0",
  "python-docx>=1.1.2",
  "python-dotenv>=1.0.1",
  "python-multipart>=0.0.9",
  "pytest>=8.3.0",
  "qdrant-client>=1.10.1",
  "sentence-transformers>=3.0.1",
  "streamlit>=1.37.0",
  "tiktoken>=0.7.0",
  "unstructured[all-docs]>=0.16.0",
  "uvicorn>=0.30.3",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
